+++ src/settings/BatchGroupConfigurationSettings/BatchGroupConfigurationForm.test.js
@@ -4,6 +4,8 @@ import { render } from '@testing-library/react';
 
 import { batchGroup, batchVoucherExport } from '../../../test/jest/fixtures';
 import {
+  EXPORT_FORMAT_OPTIONS,
+  LOCATION_TYPE_OPTIONS,
   SCHEDULE_EXPORT,
 } from './constants';
 import BatchGroupConfigurationForm from './BatchGroupConfigurationForm';
@@ -16,11 +18,14 @@ const defaultProps = {
   initialValues: {
     id: 'exportId',
     enableScheduledExport: true,
-    uploadURI: 'https://folio.com',
+    format: EXPORT_FORMAT_OPTIONS[0].value,
+    uploadURI: 'ftp.amherst-lib.edu',
+    uploadDirectory: '/files/invoices',
+    ftpFormat: LOCATION_TYPE_OPTIONS[0].value,
+    ftpPort: 22,
     scheduleExport: SCHEDULE_EXPORT.weekly,
   },
   onSubmit: jest.fn(),
-
   testConnection: jest.fn(),
   selectBatchGroup: jest.fn(),
 };